requirements.txt
click
python-dotenv
jinja2
pystackql>=3.6.1
